% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{kalshi_auth_setup}
\alias{kalshi_auth_setup}
\title{Set up Kalshi Authentication}
\usage{
kalshi_auth_setup(path = "kalshi-key-2.key")
}
\arguments{
\item{path}{Character. The absolute or relative path to your `.key` or `.pem` file.}
}
\value{
Invisibly returns the path.
}
\description{
Configures the session by pointing the package to your private RSA key.
This path is stored in an internal environment and used automatically
for all subsequent API requests.
}
\examples{
\dontrun{
kalshi_auth_setup("path/to/your/kalshi-key-2.key")
}
}
