% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{req_kalshi_authreq_kalshi_auth}
\alias{req_kalshi_authreq_kalshi_auth}
\title{Authenticate a Kalshi API Request}
\usage{
req_kalshi_authreq_kalshi_auth(req, kalshi_access_token, private_key_path)
}
\arguments{
\item{req}{An `httr2` request object.}

\item{kalshi_access_token}{Character. The Kalshi Key ID (UUID) provided in the Kalshi
web dashboard.}

\item{private_key_path}{Character. The file path to the RSA private key
(usually a .key or .pem file) used to sign the request.}
}
\value{
A modified `httr2` request object with `KALSHI-ACCESS-*` headers added.
}
\description{
This function implements the Kalshi V2 API authentication protocol. It
extracts the necessary metadata from an `httr2` request, generates a
SHA-256 RSA-PSS signature, and attaches the required headers.
}
\details{
The function automatically handles the Kalshi requirement that the signature
be generated using the base path *without* query parameters, even if the
request itself contains them.

For example, if requesting `/trade-api/v2/markets?limit=10`, this function
will correctly sign only `/trade-api/v2/markets`.
}
