% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{check_kalshi_auth}
\alias{check_kalshi_auth}
\title{Validate Kalshi Authentication Setup}
\usage{
check_kalshi_auth()
}
\value{
Returns the key path if found; otherwise, throws a cli_abort error.
}
\description{
Internal check to ensure the user has run `kalshi_auth_setup()`
before attempting an authenticated API call.
}
\keyword{internal}
